Code cleanup on aisle core.py
